+++ kernel/shared/src/main/scala/cats/effect/kernel/concurrent.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.kernel
 
 import cats.{~>, ApplicativeError, MonadError}
 import cats.syntax.either._
