+++ kernel/shared/src/main/scala/cats/effect/kernel/safe.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.kernel
 
 import cats.{ApplicativeError, MonadError}
 import cats.data.OptionT
