+++ laws/shared/src/main/scala/cats/effect/laws/AsyncLaws.scala
@@ -17,6 +17,7 @@
 package cats.effect
 package laws
 
+import cats.effect.kernel.Async
 import cats.implicits._
 
 import scala.concurrent.ExecutionContext
