+++ laws/shared/src/main/scala/cats/effect/laws/AsyncRegionLaws.scala
@@ -17,8 +17,10 @@
 package cats.effect
 package laws
 
+import cats.effect.kernel.{AsyncRegion, Bracket, Outcome}
+
 trait AsyncRegionLaws[R[_[_], _], F[_]] extends AsyncLaws[R[F, *]] with TemporalRegionLaws[R, F, Throwable] {
-  implicit val F: Async[R[F, *]] with Region[R, F, Throwable]
+  implicit val F: AsyncRegion[R, F]
 }
 
 object AsyncRegionLaws {
@@ -26,7 +28,7 @@ object AsyncRegionLaws {
       R[_[_], _],
       F[_]](
     implicit
-      F0: Async[R[F, *]] with Region[R, F, Throwable],
+      F0: AsyncRegion[R, F],
       B0: Bracket.Aux[F, Throwable, Outcome[R[F, *], Throwable, *]])
       : AsyncRegionLaws[R, F] =
     new AsyncRegionLaws[R, F] {
