+++ laws/shared/src/main/scala/cats/effect/laws/ClockTests.scala
@@ -18,6 +18,7 @@ package cats.effect
 package laws
 
 import cats.Eq
+import cats.effect.kernel.Clock
 import cats.laws.discipline._
 import cats.laws.discipline.SemigroupalTests.Isomorphisms
