+++ laws/shared/src/main/scala/cats/effect/laws/SyncLaws.scala
@@ -17,6 +17,7 @@
 package cats.effect
 package laws
 
+import cats.effect.kernel.Sync
 import cats.implicits._
 import cats.laws.MonadErrorLaws
