+++ laws/shared/src/main/scala/cats/effect/laws/TemporalRegionLaws.scala
@@ -17,8 +17,10 @@
 package cats.effect
 package laws
 
+import cats.effect.kernel.{Bracket, Outcome, TemporalRegion}
+
 trait TemporalRegionLaws[R[_[_], _], F[_], E] extends TemporalLaws[R[F, *], E] with ConcurrentRegionLaws[R, F, E] {
-  implicit val F: Temporal[R[F, *], E] with Region[R, F, E]
+  implicit val F: TemporalRegion[R, F, E]
 }
 
 object TemporalRegionLaws {
@@ -27,7 +29,7 @@ object TemporalRegionLaws {
       F[_],
       E](
     implicit
-      F0: Temporal[R[F, *], E] with Region[R, F, E],
+      F0: TemporalRegion[R, F, E],
       B0: Bracket.Aux[F, E, Outcome[R[F, *], E, *]])
       : TemporalRegionLaws[R, F, E] =
     new TemporalRegionLaws[R, F, E] {
