+++ laws/shared/src/test/scala/cats/effect/OutcomeSpec.scala
@@ -17,6 +17,8 @@
 package cats.effect
 
 import cats.{Eval/*, Id*/}
+import cats.effect.kernel.Outcome
+import cats.effect.testkit.OutcomeGenerators
 import cats.implicits._
 import cats.laws.discipline.{ApplicativeErrorTests, MonadErrorTests}
