+++ laws/shared/src/test/scala/cats/effect/TimeTSpec.scala
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.testkit
 
 import cats.{Eq, Order}
 import cats.data.Kleisli
 import cats.effect.laws.TemporalBracketTests
+import cats.effect.testkit.{pure, OutcomeGenerators, PureConcGenerators, TimeT}, pure.PureConc
 import cats.implicits._
 import cats.laws.discipline.arbitrary._
 
-import pure._
 import TimeT._
 
 import org.specs2.ScalaCheck
