+++ testkit/shared/src/main/scala/cats/effect/testkit/FreeSyncGenerators.scala
@@ -15,12 +15,12 @@
  */
 
 package cats.effect
-package laws
+package testkit
 
 import cats.{Eval, Monad, MonadError}
+import cats.effect.kernel.Sync
+import cats.effect.testkit.freeEval._
 import cats.free.FreeT
-
-import freeEval._
 
 import org.scalacheck.{Arbitrary, Cogen, Gen}
